refactor: into modfile and module
